# PERF-#5778: Avoid extra materialization at range-based reshuffling